requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
db-dtypes
pyarrow
